backend/requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
pydantic==2.12.5
pydantic_core==2.41.5
pydantic-settings==2.12.0
python-dotenv==1.2.1
starlette==0.50.0
bcrypt==5.0.0

pydantic[email]
python-jose[cryptography]
sqlalchemy[asyncio]
asyncpg
alembic
psycopg[binary]

aiofiles
graypy
fast-graphrag>=0.0.5
openai>=1.10.0
instructor>=1.0.0
google-genai>=1.0.0

python-multipart>=0.0.6
